Update functions related FEniCSx

+++ single_run/structureDomain/structureFSISetup/structureBCS.py
@@ -38,7 +38,7 @@ __license__ = "All rights reserved"
 #%% Import packages
 #_________________________________________________________________________________________
 
-from dolfin import *
+from dolfinx import *
 import numpy as np
 
 #_________________________________________________________________________________________
@@ -53,7 +53,7 @@ class boundaryConditions:
         bc2 = DirichletBC(MixedVectorFunctionSpace.sub(1), ((0.0,0.0,0.0)),boundaries, marks)
         return bc1, bc2
     def DirichletBCs(self, VectorFunctionSpace, boundary_dofs):
-        bc3 = fem.dirichletbc(np.zeros(3, dtype=dtype),
+        bc3 = fem.dirichletbc(np.zeros(3),
                               boundary_dofs,
                               V=VectorFunctionSpace)
         return bc3
